[eve MAINLINE] Fix the manifests.props chaos
Ok, so, I must have had a stroke because I thought these files were the same, they are not. So I introduced the old one in a new place that works for what seems to be non ccpproj setups. Although ExeFile seems to be half converted to ccpproj, so fuck him, he's getting the old one.

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Exefile/": change = 1214960]