[eve MAINLINE] Send minidumps to sentry.io
exeFile.exe now sends to both CCP crashes and sentry.io

The curent breakpad version seems to have some limitation in the size of headers which can be sent with the minidump so we just decided on the most important tags for now. We can add more, just not sure how much more.

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Exefile/": change = 1487998]