Stop importing blue from exefile
Importing blue from exefile causes a double import of the module in non-release builds. This is because `LoadPythonExtension` does not add `blue` into `sys.modules`, but rather `blue<_flavor>`. Additionally, because `blue` is correctly wired up to Python's builtin modules list via extending the so called "inittab", then a `LoadPythonExtension("blue")` call to load the correct flavor of blue is no longer needed.