Break circular dependency between exefile and blue
Exefile used to require some headers from blue to accommodate for the `IBeOS` and `IBluePaths` interfaces. However, this meant that exefile needed to include blue as a dependency, even if it was just to locate the headers.
However, exefile is now bootstrapping so much of the process environment that blue needs to run its own tests through exefile. Therefore, relying on the blue package is no longer an option for exefile. Fortunately, blue was already lazy-loaded at runtime, and therefore untangling the dependency is mostly a case of changing the required C API exposed by blue.